% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIMMSC_data.R
\docType{data}
\name{data_DIMMSC}
\alias{data_DIMMSC}
\title{A simulated droplet-based single cell data}
\format{An object of class \code{matrix} with 1000 rows and 6000 columns.}
\usage{
data("data_DIMMSC")
}
\description{
A simulated droplet-based single cell data
}
\examples{
# Load the example data data_DIMMSC
data("data_DIMMSC")

# Run DIMMSC
result <- DIMMSC(data=data_DIMMSC, K=3, method_cluster_initial="kmeans",
                 method_alpha_initial="Ronning", maxiter=200, tol=1e-4, lik.tol=1e-2)

# Plot t-SNE and clusters
plot_tsne_clusters(data=data_DIMMSC, cluster=result$mem)
}
\keyword{datasets}
