% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIMMSC.R
\name{DIMMSC}
\alias{DIMMSC}
\alias{plot_tsne_clusters}
\title{Clustering droplet-based single cell transcriptomic data using Dirichlet mixture prior and EM algorithm}
\usage{
DIMMSC(data, K = 2, method_cluster_initial = "kmeans",
  method_alpha_initial = "Ronning", maxiter = 200, tol = 1e-04,
  lik.tol = 0.01, customized_initial = NULL)

plot_tsne_clusters(data, cluster)
}
\arguments{
\item{data}{a G*C matrix with G genes and C cells}

\item{K}{number of desired clusters}

\item{method_cluster_initial}{method for initializing clusters, "kmeans", "random" or "customized"}

\item{method_alpha_initial}{method for initializing the alpha matrix for EM algorithm, "Ronning" (Ronning's method, 1989) or "Weir" (Weir and Hill's method 2002)}

\item{maxiter}{maximum number of iterations}

\item{tol}{a convergence tolerance for the difference of vector pie between iterations}

\item{lik.tol}{a convergence tolerance for the difference of log-likelihoods between iterations}

\item{customized_initial}{a vector of positive integers \{1,2,...,K\} for the customized initial clusters, the length should be equal to the number of cells}

\item{cluster}{a vector of clustering member ship, e.g. mem of DIMMSC output}
}
\value{
DIMMSC returns a list object containing:

\itemize{
  \item pie: a vector of pie estimates
  \item delta: a C*K matrix with probability that each cell belongs to each cluster
  \item alpha: a K*G matrix of alpha estimates
  \item mem: a vector of clustering label
  \item loglik: the final log likelihood after iterations
  \item AIC: Akaike information criterion (AIC)
  \item BIC: Bayesian information criterion (BIC)
}

plot_tsne_clusters outputs a figure of t-SNE and clusters
}
\description{
DIMMSC is for clustering droplet-based single cell transcriptomic data. It uses Dirichlet mixture prior to characterize variations across different clusters. An EM algorithm is used for parameter inference. This package can provide clustering uncertainty.

plot_tsne_clusters performs log2 normalization, PCA and t-SNE on the data and plot clusters on the t-SNE projection
}
\examples{
# Load the example data data_DIMMSC
data("data_DIMMSC")

# Run DIMMSC
result <- DIMMSC(data=data_DIMMSC, K=3, method_cluster_initial="kmeans",
                 method_alpha_initial="Ronning", maxiter=200, tol=1e-4, lik.tol=1e-2)

# Plot t-SNE and clusters
plot_tsne_clusters(data=data_DIMMSC, cluster=result$mem)
}
\references{
Zhe Sun, Ting Wang, Ke Deng, Xiao-Feng Wang, Robert Lafyatis, Ying Ding, Ming Hu, Wei Chen. DIMM-SC: A Dirichlet mixture model for clustering droplet-based single cell transcriptomic data. Bioinformatics 2017.
}
\author{
Zhe Sun <zhs31@pitt.edu>, Ting Wang <tiw33@pitt.edu>, Ark Fang <zhf9@pitt.edu>, Wei Chen <wei.chen@chp.edu>.
}
